+++ src/GRA.Data/Repository/SystemRepository.cs
@@ -1,10 +1,12 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
+﻿using AutoMapper.QueryableExtensions;
+using GRA.Domain.Model.Filters;
 using GRA.Domain.Repository;
-using Microsoft.Extensions.Logging;
+using GRA.Domain.Repository.Extensions;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
+using System.Collections.Generic;
 using System.Linq;
-using AutoMapper.QueryableExtensions;
+using System.Threading.Tasks;
 
 namespace GRA.Data.Repository
 {
@@ -23,6 +25,40 @@ namespace GRA.Data.Repository
                 .OrderBy(_ => _.Name)
                 .ProjectTo<Domain.Model.System>()
                 .ToListAsync();
+        }
+
+        public async Task<int> CountAsync(BaseFilter filter)
+        {
+            return await ApplyFilters(filter)
+                .CountAsync();
+        }
+
+        public async Task<ICollection<Domain.Model.System>> PageAsync(BaseFilter filter)
+        {
+            return await ApplyFilters(filter)
+                .OrderBy(_ => _.Name)
+                .ApplyPagination(filter)
+                .ProjectTo<Domain.Model.System>()
+                .ToListAsync();
+        }
+
+        private IQueryable<Data.Model.System> ApplyFilters(BaseFilter filter)
+        {
+            var systemList = DbSet
+                 .AsNoTracking()
+                 .Where(_ => _.SiteId == filter.SiteId);
+
+            if (!string.IsNullOrWhiteSpace(filter.Search))
+            {
+                systemList = systemList.Where(_ => _.Name.Contains(filter.Search));
+            }
+
+            return systemList;
+        }
+
+        public async Task<bool> IsInUseAsync(int systemId)
+        {
+            return await _context.Branches.AsNoTracking().AnyAsync(_ => _.SystemId == systemId);
         }
 
         public async Task<bool> ValidateAsync(int systemId, int siteId)