+++ src/GRA.Domain.Model/Permission.cs
@@ -26,6 +26,7 @@ namespace GRA.Domain.Model
         ManageLocations,
         ManageQuestionnaires,
         ManageSchools,
+        ManageSystems,
         ManageTriggers,
         ManageTriggerMail,
         ManageVendorCodes,