+++ src/GRA.Domain.Repository/IBranchRepository.cs
@@ -1,4 +1,5 @@
 ﻿using GRA.Domain.Model;
+using GRA.Domain.Model.Filters;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -8,6 +9,10 @@ namespace GRA.Domain.Repository
     {
         Task<IEnumerable<Branch>> GetAllAsync(int siteId);
         Task<IEnumerable<Branch>> GetBySystemAsync(int systemId);
+        Task<ICollection<Branch>> PageAsync(BaseFilter filter);
+        Task<int> CountAsync(BaseFilter filter);
+        Task<bool> IsInUseAsync(int branchId);
+
         Task<bool> ValidateAsync(int branchId, int systemId);
         Task<bool> ValidateBySiteAsync(int branchId, int siteId);
     }