+++ src/GRA.Domain.Repository/ISystemRepository.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using GRA.Domain.Model.Filters;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace GRA.Domain.Repository
@@ -6,6 +7,9 @@ namespace GRA.Domain.Repository
     public interface ISystemRepository : IRepository<Model.System>
     {
         Task<IEnumerable<Model.System>> GetAllAsync(int siteId);
+        Task<ICollection<Model.System>> PageAsync(BaseFilter filter);
+        Task<int> CountAsync(BaseFilter filter);
+        Task<bool> IsInUseAsync(int systemId);
         Task<bool> ValidateAsync(int systemId, int siteId);
     }
 }